renamed GamecontentsPicker to GameComponents, where all stuff with components will be moved
