+++ client/src/GameComponents/GameVersionPicker.js
@@ -16,7 +16,7 @@ export const GameVersion = {
 // WARNING if you're changing this default version, change the default in /server/Startup.cs
 export const DEFAULT_VERSION = GameVersion.PoK_Codex2
 
-export const GameContentsPicker = ({ value, onChange }) => {
+export const GameVersionPicker = ({ value, onChange }) => {
   const getFromEvent = (event) => {
     const v = event.target.value
     onChange(Number(v))
