+++ client/src/GameComponents/index.js
@@ -1,2 +1,2 @@
-export { GameVersion, GameContentsPicker } from './GameContentsPicker'
+export { GameVersion, GameVersionPicker } from './GameVersionPicker'
 export { useGameVersion, GameVersionProvider } from './useGameVersion'
